Small Bugfixes (#276)
* textureCache writing

- texture cache writing and creation from dds files
- more options for Wkit.Console cache

* cubemap support

* Serialization Versions for ArchiveManagers

- added serialization versions to all serialized ArchiveManagers used for updating ManagerCaches in case the class changes

* bug fixes

- fixed a bug where uncooking files would uncook too many files
- fixed soe UI issues

+++ WolvenKit.W3Speech/DataStructures.cs
@@ -63,7 +63,7 @@ namespace WolvenKit.W3Speech
 
         public IWitcherArchive Bundle { get; set; }
         public string Name { get; set; }
-        public long Size { get; set; }
+        public uint Size { get; set; }
         public uint ZSize { get; set; }
 
         public long PageOffset { get; set; }
